app/requirements.txt
anyio==4.13.0
click==8.3.1
grpcio==1.80.0
grpcio-tools==1.80.0
h11==0.16.0
idna==3.11
protobuf==6.33.6
python-multipart==0.0.22
redis==7.4.0
setuptools==82.0.1
starlette==1.0.0
typing_extensions==4.15.0
uvicorn==0.42.0
